+++ app/src/main/java/com/lop/budget/data/local/entity/TransactionEntity.kt
@@ -33,7 +33,8 @@ import com.lop.budget.domain.model.TransactionType
 data class TransactionEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val title: String,
-    val amount: Double,
+    /** Montant en centimes, toujours positif : le sens est porté par [type]. */
+    val amount: Long,
     val type: TransactionType,
     val status: TransactionStatus,
     val kind: TransactionKind = TransactionKind.STANDARD,
