+++ app/src/main/java/com/lop/budget/domain/model/Balance.kt
@@ -2,4 +2,5 @@ package com.lop.budget.domain.model
 
 import com.lop.budget.data.local.entity.AccountEntity
 
-data class AccountBalance(val account: AccountEntity, val balance: Double)
+/** [balance] est exprimé en centimes. */
+data class AccountBalance(val account: AccountEntity, val balance: Long)
